expose interfaces
